apps/api/requirements.txt
fastapi
uvicorn
ruff
sqlalchemy
pydantic-settings
python-dotenv
loguru
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt
